design: add direct link to discussion in template
Traditionally, Go design docs linked to the corresponding mailing list
thread for discussion. In the new world, discussion happens on the
GitHub issue. This is easier to get to because the name of the design
doc implies the GitHub issue, but this requires both knowledge of this
convention and additional effort on the reader's part.

Add a direct link to the discussion in the design doc template to make
it easy to find the discussion with no additional knowledge or effort.

Change-Id: Ie76dd85551bc33dff7d981fbd0a28e664947b005
Reviewed-on: https://go-review.googlesource.com/15290
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>