12416: write rules in terms of an invariant
Using an invariant makes this easier to understand.

Drop the idea of a syntactic checker in cmd/cgo.  Since it doesn't match
the rules or the invariant, and is not a superset either, it just
becomes another thing to learn rather than being helpful.  Instead,
describe a dynamic checker that enforces the invariant.

Change-Id: I2309d118020e2d10af73e1e33ea8c71fe32a375a
Reviewed-on: https://go-review.googlesource.com/15831
Reviewed-by: Andrew Gerrand <adg@golang.org>