proposal: Modify table proposal to add down-slicing and reshaping.
This change modifies the table proposal to allow down-slicing and reshaping,
and while here fixes some spelling mistakes. Down-slicing allows for a
portion of a higher-dimensional table to be seen as a lower-dimensional
one, and reshaping allows for a linear array to be seen as a higher-dimensional
one. A consequence of these changes is that the proposed implementation of
copy is simplified significantly.

Updates golang/go#6282.

Change-Id: I8bb2376c4f12c1fde695d23853734841abfdcdbf
Reviewed-on: https://go-review.googlesource.com/24271
Reviewed-by: Andrew Gerrand <adg@golang.org>
Reviewed-by: Robert Griesemer <gri@golang.org>