design/14313: reconcile with testing package format
The benchmark format proposal was written in parallel with the
sub-benchmarks support and the two didn't quite agree on the format of
benchmark names.

This commit makes two changes to the benchmark format document so that
it agrees with the practice and implementation of sub-benchmarks in
the testing package:

* It replaces key:value in the benchmark name with key=value, which is
  the format suggested by the sub-benchmarks documentation.

* It removes the special rule about reporting gomaxprocs as a special
  key=value pair. This was never implemented in the testing package,
  so it still emits a -N suffix for the gomaxprocs value. Technically
  this syntax is ambiguous, but it's what we're doing.

Updates golang/go#14313.

Change-Id: I67cbd735de7ab145bc7c3a93b5dadbe05120929c
Reviewed-on: https://go-review.googlesource.com/33570
Reviewed-by: Marcel van Lohuizen <mpvl@golang.org>