proposal: small update to DWARF inlining proposal
Update the proposal document to reflect the fact that DW_AT_call_file
and DW_AT_call_line attributes are required for any generated
DW_TAG_inlined_subroutine DIEs. This isn't spelled out as required in
the standard, but it seems to be enforced by GDB, so the compiler will
need to emit these.

Updates golang/go#22080

Change-Id: I467773580ceff7ce4d36ec919685909d43e8addf
Reviewed-on: https://go-review.googlesource.com/72750
Reviewed-by: Austin Clements <austin@google.com>