design/24543: split out safe-points-everywhere
The non-cooperative preemption proposal document currently mixes
discussion of non-cooperative preemption, which is a very general
idea, with a specific implementation based on stack maps and register
maps at every instruction.

I'm about to introduce a proposal for an alternative implementation
approach to non-cooperative preemption. In order to prepare for that,
this CL splits the existing document into a general overview of
non-cooperative preemption, and a separate document detailing the
safe-points-everywhere implementation approach. This is similar to how
we organize the generics proposal (15292-generics.md).

Updates golang/go#24543.

Change-Id: Iacd1cff3f00ec9e153122d5eb21a9a1699630e34
Reviewed-on: https://go-review.googlesource.com/c/158859
Reviewed-by: Michael Knyszek <mknyszek@google.com>