design/19308-number-literals: adjust hex_float_lit and imaginary_lit syntax
We permit a _ separator after the hex prefix in a hexadecimal
integer, as in 0x_1234; for consistency we also allow the
separator in a hexadecimal float, as in 0x_1234.0p0.

Also, to limit imaginary literals to decimal representations
as is the case now, change the repespective syntax to refer
to decimal_float_lit rather than float_lit (which now also
permits hexadecimal float representations).

Change-Id: I03d95dffe3d583dad862b4f75e998aa0772f39a2
Reviewed-on: https://go-review.googlesource.com/c/158862
Reviewed-by: Russ Cox <rsc@golang.org>