design/35112-scaling-the-page-allocator.md: update proposal
This change updates the scaling-the-page-allocator proposal to better
reflect the final implementation that ended up landing. It also updates
the "Prior Art" section with references to DieHard, DieHarder, and
Mesh, which are malloc/free implementations that notably use bitmaps as
a part of their implementation.

Updates golang/go#35112.

Change-Id: If584fca2a3928e3c7518d0a8df806809ca69c5e2
Reviewed-on: https://go-review.googlesource.com/c/proposal/+/219121
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Trust: Michael Knyszek <mknyszek@google.com>