14313-benchmark-format: minor refreshes
This updates some links and some stale information in this design doc.

This is slightly odd because we use this design doc also as a
specification, which creates some tension between whether it should
continue to capture the state of the world at the time or be kept
fully up-to-date. I tried to take a light touch in this CL.

Change-Id: Ifa3cf7d70e0f6a65cd9184f76f449d5618211f63
Reviewed-on: https://go-review.googlesource.com/c/proposal/+/357529
Reviewed-by: Michael Knyszek <mknyszek@google.com>