design/56345-structured-logging.md: add design doc
This design document is a modification of the top post of the GitHub
discussion on structured logging
(https://github.com/golang/go/discussions/54763).

Most of the document is unchanged, except for reformatting.
The section on contexts has been rewritten to accommodate the new
`WithContext` method and `Ctx` function, and the sections on groups
and the `LogValue` method have been added.

Change-Id: Ife6385c87c872c2a31c036ca63f8f61e6e2902a0
Reviewed-on: https://go-review.googlesource.com/c/proposal/+/444415
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Russ Cox <rsc@golang.org>