README: add Go architects
Add "Go architects" to handle cases where proposal review cannot
identify a consensus in the issue discussion.

The proposal review group reviews proposals weekly to help them
through the process. (See
https://github.com/golang/proposal#proposal-review.)

When a proposal does not have a clear outcome but a decision must be
made, the proposal review group also works to reach consensus among
themselves on the answer; that is, they decide the outcome. (See
https://github.com/golang/proposal#consensus-and-disagreement.)

Reviewing proposals and deciding contended proposals are two different
activities that need not be done by exactly the same set of people.
Historically there was significant overlap, and there will likely
always be some overlap, but as time goes on it seems likely that we
may want to formally separate the two activities.

In particular, we may want to add more reviewers, to help the process
move along and scale, but still leave truly contended decisions (which
are fairly rare) to a smaller group. This change to README.md implements
that idea.

Fixes golang/go#33528.

Change-Id: Iaec6d7fc4ff588f057eb5db84a9de548ec4c2b8d
Reviewed-on: https://go-review.googlesource.com/c/proposal/+/506565
Reviewed-by: Russ Cox <rsc@golang.org>
TryBot-Bypass: Russ Cox <rsc@golang.org>